ABI: Remove work item topics for standardizing psABIs
As far as I know, there is no precedence for standardizing on psABIs.
For example, there is no standard for x86-64 calling conventions, nor
are there any standards extensions for ELF extensions for x86-64,
aarch64, SPARC, MIPS, etc.

As such unless we have a very good reason for diverging from a very
widely accepted industry norm, we should not be the first organization
to standardize on psABIs. Let's remove this from the working group
charter. We can always standardize it in subsequent working groups if
it becomes clear that it's necessary.

Signed-off-by: David Vernet <void@manifault.com>